overhaul of files
I realized that I needed to really do a thorough check and correction of the address data. I was initially looking for duplicate addresses (people who hadn't moved but appeared to because the lat/long was slightly off). I then realized that there were a lot of addresses that had lat/long switched. I manually corrected these and saved that in geocodes_corrected.csv on my desktop - I'll move that to the SHAG folder next. I also created geocodes_PEGS.R exclusively to follow that process. caces_pegs.R is exclusively to tie the PEGS addresses to the CACES dataset.